# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Seed-IOT",
        "title": "account api",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
        "/account/bindings": {
            "post": {
                "description": "绑定设备",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "BindDevices",
                "operationId": "BindDevices",
                "parameters": [
                    {
                        "description": "用户登录",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/web.BindDevicesParames"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/constant.BaseReturn"
                        }
                    }
                }
            }
        },
        "/account/getCode": {
            "post": {
                "description": "获取验证码",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Get Code",
                "operationId": "Get Code",
                "parameters": [
                    {
                        "description": "获取验证码",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/web.GetCodeParam"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/constant.BaseReturn"
                        }
                    }
                }
            }
        },
        "/account/login": {
            "post": {
                "description": "账号密码登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Login",
                "operationId": "Login",
                "parameters": [
                    {
                        "description": "用户登录",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/web.LoginParames"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/constant.BaseReturn"
                        }
                    }
                }
            }
        },
        "/account/mobileLogin": {
            "post": {
                "description": "验证码登录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Mobile Login",
                "operationId": "Mobile Login",
                "parameters": [
                    {
                        "description": "用户登录",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/web.MobileLoginParames"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "schema": {
                            "$ref": "#/definitions/constant.BaseReturn"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "constant.BaseReturn": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "web.BindDevicesParames": {
            "type": "object",
            "required": [
                "snos",
                "uid"
            ],
            "properties": {
                "snos": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "uid": {
                    "type": "string"
                }
            }
        },
        "web.GetCodeParam": {
            "type": "object",
            "required": [
                "codeType",
                "mobile"
            ],
            "properties": {
                "codeType": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                }
            }
        },
        "web.LoginParames": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "web.MobileLoginParames": {
            "type": "object",
            "required": [
                "code",
                "mobile"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                }
            }
        }
    }
}